Extract name/type/expr conversions to separate files
